Added the Control template class
The Control class is a base class to make new UI controls. It owns the
main common features that a control can offer (e.g., a background, a
content item, etc.). In addition, the Control class comes with a padding
property so to add spaces between the control and its content item.

Finally, in the tests projects, conditional properties have been added for enabling
or disabling the tests and autotests projects. Those variables must be set
manually in the top-level project.