Separated StoiridhControlsTemplates in two API
StoiridhControlsTemplates comes now with both a private and a public API by
using the d-pointer pattern provided (internally) by Qt. It is necessary to
provide a binary compatibility for this project. This implies the use of
the Qt private API, specifically the use of Qt Core, Qt GUI, Qt Qml, and
Qt Quick private API. For now, the use of the Qt private API is limited to
the d-pointer pattern.

StoiridhControlsTemplates:

In the Control class, the properties 'Background' and 'Content' are now
designable.