Document Control and Padding classes
Sphinx provides a C++ extension in order to document the C++ code. But
it is not suited for documenting Qt code, e.g., signals, properties, and
slots. So to bypass those limitations, I split the documentation of the
C++ functions into sub-sections.

note: the exceptions are deliberately not documented for now.