Migrate the documentation to Sphinx (#39)
* Move the FDL licence

The FDL licence file has been moved at the root of the project.

* Update the CI configuration file

The APT repository for Qt has been updated to Qt 5.6.1 as well as its
repository name.

* Remove the Stòiridh QDoc Template submodule

This (deprecated) submodule is no longer required since we'll migrate to
Sphinx.

* Ignore the build directory

Sphinx will generate a directory named build before starting to build
the documentation.

* Remove the Qt documentation

All files related to Qt documentation is removed in order to create the
Sphinx documentation.

* Add the Sphinx project

Plus a short introduction of the project in the main page.

* Document Control and Padding classes

Sphinx provides a C++ extension in order to document the C++ code. But
it is not suited for documenting Qt code, e.g., signals, properties, and
slots. So to bypass those limitations, I split the documentation of the
C++ functions into sub-sections.

note: the exceptions are deliberately not documented for now.

Task-number: #38 
Close: #38